添加随机吃什么功能，显示可以抽取类型的功能，emojimix功能

+++ src/main/kotlin/entity/User.kt
@@ -13,6 +13,6 @@ class User(val account: Member) {
         }
     }
 
-    var luckyValue: Int = -1
+    var luckyValue: Double = -1.0
     var luckyValueAcquisitionDate: String = SimpleDateFormat("yyyy/MM/dd").format(Date())
 }
